Added more debugging support

+++ src/dtype/fdtype.c
@@ -112,7 +112,8 @@ FD_EXPORT void fd_init_pptrs_c(void);
 FD_EXPORT void fd_boot_message()
 {
   u8_message("Copyright (C) beingmeta 2004-2016, all rights reserved");
-  u8_message("(%s:%ld) %s %s",u8_appid(),getpid(),
+  u8_message("(%s:%lld) %s %s",
+             u8_appid(),(unsigned long long)getpid(),
              fd_getrevision(),u8_getrevision());
 }
